feat: Implement Stripe payment integration and enhance order delivery features
- Added StripeCallbackController to handle success and refresh callbacks for Stripe onboarding.
- Updated StripeConnectController to improve account connection logic and onboarding link generation.
- Introduced CreateWithdrawalRequest for validating withdrawal requests with balance checks.
- Enhanced SubmitDeliveryRequest to allow larger message and file sizes.
- Added activities relationship to Order model for tracking order activities.
- Updated OrderDelivery model to include QA feedback and review timestamps.
- Improved Room model methods for better readability and functionality.
- Refactored InboxService to utilize new getOtherUserAttribute method.
- Updated QaService to streamline delivery approval and rejection processes.
- Enhanced StripePaymentService for better account management and checkout session creation.
- Updated configuration for order settings to ensure proper type casting.
- Made UI adjustments in QA views for better layout and user experience.
- Updated API routes for clearer structure and improved endpoint organization.

+++ config/orders.php
@@ -3,7 +3,7 @@
 return [
     // 'auto_complete_days' => env('ORDER_AUTO_COMPLETE_DAYS', 7),
     'auto_complete_days' => (int) env('ORDER_AUTO_COMPLETE_DAYS', 7),
-    'auto_accept_days'   => env('ORDER_AUTO_ACCEPT_DAYS', 3),
-    'escrow_hold_days'   => env('ORDER_ESCROW_HOLD_DAYS', 14),
-    'platform_fee_pct'   => env('ORDER_PLATFORM_FEE_PCT', 10),
+    'auto_accept_days'   => (int) env('ORDER_AUTO_ACCEPT_DAYS', 3),
+    'escrow_hold_days'   => (int) env('ORDER_ESCROW_HOLD_DAYS', 14),
+    'platform_fee_pct'   => (int) env('ORDER_PLATFORM_FEE_PCT', 10),
 ];
